Fix mock_payment_method's `last4` extraction
Previously, passing in a params hash with an integer card number to
`StripeMock::Data.mock_payment_method` would raise
`TypeError: no implicit conversion of Range into Integer`. This is the
case for a couple of shared RSpec examples.

This change ensures the "dug" value is always converted to a String
before attempting to index it.

Also provide an end to the ranges in StripeMock::Data methods to
support the Ruby 2.4.0 requirement listed in the README. (see
https://bugs.ruby-lang.org/issues/12912.)